Extract common test functions
